refactor: 메일 API 변경 (#114)
